Assignment-5 Personal AI Assistant (Gemini API + OOP + Streamlit)/requirements.txt
google-generativeai 
python-dotenv 
streamlit
SpeechRecognition
gTTS
